Resolve Wallpaper | Configure issues
On Samsung Galaxy S (and possibly other) phones, the custom skin has
a Menu | Wallpaper | Configure menu which is a shortcut for launching
the LWP preferences.

When the activity uses Theme.WallpaperSettings, launching preferences via
Configure results in onSurfaceChanged NOT being called. The LWP continues
to run (the activity is transparent), and we needed to listen for changes
in the preferences so that we could update the running LWP. This led to
race conditions and other badness that caused crashes.

When preferences are launched from Settings, the LWP is paused (we can
see this through the transparent activity). After preferences activity
is dismissed, the LWP is resumed and we grab the new settings and update
the LWP.

By removing the theme, the activity is opaque, but in both launch cases,
the LWP is paused, which is what we want.