Another attempt at fixing pulsing bug
When pulsing happens, # of particles alive == max particles. So no
particles are spawned for that frame, but the # due to be spawned
increases. So the trick is to prevent max particles from ever being
spawned. Change the max rate so that the oldest particles will expire
before we ever reach maxParticles.