Remove unnecessary shuffling
This prevents dots from overlapping, since shuffling could cause the
position of a live particle to be re-used before the particle has
expired. It means that the particles will always be drawn to the same
positions in the same order, but each frame spawns a different number
of particles, so a pattern shouldn't be apparent.